% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_maf.R
\name{convert_vcf_to_maf}
\alias{convert_vcf_to_maf}
\title{VCF2MAF}
\usage{
convert_vcf_to_maf(path_to_vcf)
}
\arguments{
\item{path_to_vcf}{}
}
\value{
MAF dataframe
}
\description{
Convert a vcf (or vcfs) to a MAF file
}
\examples{

}
